Move socket
