+++ collab-service/package.json
@@ -12,5 +12,9 @@
   "dependencies": {
     "express": "^4.18.1",
     "socket.io": "^4.5.2"
+  },
+  "optionalDependencies": {
+    "bufferutil": "^4.0.6",
+    "utf-8-validate": "^5.0.9"
   }
 }
