+++ frontend/src/components/editor/CodeEditorWindow.js
@@ -1,10 +1,8 @@
-import React, { useEffect, useState } from "react";
+import React, {  useState } from "react";
 import { useLocation } from "react-router-dom";
-import { io } from "socket.io-client";
 import Editor from "@monaco-editor/react";
-import { PREFIX_COLLAB_SVC, URL_COLLAB_SVC } from "../../configs";
 
-const CodeEditorWindow = ({ onChange, language, code, theme }) => {
+const CodeEditorWindow = ({ onChange, language, code, theme, socket }) => {
   const location = useLocation(); // Location contains username and selected difficulty level
   const [value, setValue] = useState(code || "");
 
@@ -20,21 +18,6 @@ const CodeEditorWindow = ({ onChange, language, code, theme }) => {
     })
     updateValue(value);
   };
-
-  const socket = io(URL_COLLAB_SVC, { 
-    transports: ['websocket'],
-    path: PREFIX_COLLAB_SVC
-  });
-
-  useEffect(() => {
-    console.log(`finding room_id=${location.state.room_id}`);
-    // Emit matching event here
-    socket.emit('room', { room_id: location.state.room_id });  
-
-    return () => { // component will unmount equivalent
-      socket.emit('leave room', { room_id: location.state.room_id });
-    }
-  }, [])
 
   socket.on('receive code', (payload) => {
     console.log(`received payload ${payload.newCode}`)
